COMMON, LAUNCHER, SERVICE: Moved the Manifests and Messaging solutions into a single merged solution, CoffeeTable.Common, and merged the namespaces into CoffeeTable.Common.Manifests and CoffeeTable.Common.Messaging respectively.
This makes the process of importing the shared coffee table files easier between applications, and provides a single place to keep shared files. Instead of linking the files in visual studio (which can be annoying), use the built CoffeeTable.Common dll. Also fixed a bug in the launcher where all directories were searched recursively in the apps folder instead of only searching the topmost directory